+++ scripts/lichessTools-init.js
@@ -60,6 +60,8 @@ lichessTools.loadTool(LiChessTools.Tools.ChapterInsert);
 lichessTools.loadTool(LiChessTools.Tools.ChapterPgnArea);
 lichessTools.loadTool(LiChessTools.Tools.MoveListOptions);
 lichessTools.loadTool(LiChessTools.Tools.StreamerTv);
+lichessTools.loadTool(LiChessTools.Tools.WikiFen);
+lichessTools.loadTool(LiChessTools.Tools.TranspositionArrows);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);
 lichessTools.loadTool(LiChessTools.Tools.TrapValueCommand);