+++ scripts/lichessTools.js
@@ -312,7 +312,7 @@
       return this.getElementForPath(path);
     }
 
-    traverse=(snode, func)=>{
+    traverse=(snode, func, forced)=>{
       if (!snode) {
         snode=this.lichess?.analysis?.tree.root;
         this.isTreeviewVisible(true);
@@ -332,7 +332,7 @@
       }];
       while (nodes.length) {
         let {node,path}={...nodes.shift()};
-        if (!this.isTreeviewVisible()) return;
+        if (!forced && !this.isTreeviewVisible()) return;
         if (!node || node.comp) {
           continue;
         }
@@ -680,7 +680,7 @@
       const age=lichess.info?.date
         ? (new Date().getTime()-new Date(lichess.info.date).getTime())/86400000
         : 0;
-      console.debug('%c lichess.org code age: '+Math.round(age*10)/10+' days', age<7?'background: red;':'');
+      console.debug('%c site code age: '+Math.round(age*10)/10+' days', age<7?'background: red;':'');
       this.translator = this.lichess.trans(this.intl.siteI18n);
       await this.applyOptions();
       const debouncedApplyOptions=this.debounce(this.applyOptions,250);