+++ HISTORY.md
@@ -1,5 +1,14 @@
 History of features added to LiChess Tools in time (reversed order)
 
+18 Nov
+  - Interactive lesson Preview mode sticks between page reloads as well
+  - Option to open links from move list comments that lead to studies in the same window
+  - Transpositions in variation arrows
+
+16 Nov
+  - Analysis Wiki FEN search
+  - Copy PGN tohere parameter
+
 13 Nov
   - Alt-click to copy PGN until current move
   - added Move eval button to the right to move the computer evaluation toggle back to the right side